requirements.txt
Django==4.2.10
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-cors-headers==4.3.1
Pillow==10.1.0
minio==7.2.0
gunicorn==21.2.0
django-storages==1.14.2
boto3==1.28.65